goalBased/requirements.txt
langchain==0.2.14
langchain-core==0.2.33
langchain-google-genai==1.0.8
streamlit
python-dotenv
PyMuPDF

#pip langchain-community
# langchain-openai
# openai